+++ lending/CodatLending/Models/Operations/GetAccountingBalanceSheetResponse.cs
@@ -11,7 +11,6 @@
 namespace CodatLending.Models.Operations
 {
     using CodatLending.Models.Shared;
-    using System.Collections.Generic;
     using System.Net.Http;
     using System;
     
@@ -21,7 +20,7 @@ namespace CodatLending.Models.Operations
         /// <summary>
         /// Success
         /// </summary>
-        public Dictionary<string, object>? AccountingBalanceSheet { get; set; }
+        public AccountingBalanceSheet? AccountingBalanceSheet { get; set; }
 
         /// <summary>
         /// HTTP response content type for this operation