+++ lending/CodatLending/Models/Shared/BankingAccount.cs
@@ -11,7 +11,6 @@
 namespace CodatLending.Models.Shared
 {
     using Newtonsoft.Json;
-    using System.Collections.Generic;
     
     /// <summary>
     /// This data type provides a list of all the SMB&apos;s bank accounts, with rich data like balances, account numbers, and institutions holding the accounts.<br/>
@@ -30,7 +29,7 @@ namespace CodatLending.Models.Shared
         /// Depending on the data provided by the underlying bank, not all balances are always available.
         /// </summary>
         [JsonProperty("balance")]
-        public Dictionary<string, object> Balance { get; set; } = default!;
+        public AccountBalanceAmounts Balance { get; set; } = default!;
 
         /// <summary>
         /// The currency code for the account.
@@ -54,7 +53,7 @@ namespace CodatLending.Models.Shared
         /// An object containing bank account identification information.
         /// </summary>
         [JsonProperty("identifiers")]
-        public Dictionary<string, object> Identifiers { get; set; } = default!;
+        public AccountIdentifiers Identifiers { get; set; } = default!;
 
         /// <summary>
         /// The friendly name of the account, chosen by the holder. This may not have been set by the account holder and therefore is not always available.
@@ -66,7 +65,7 @@ namespace CodatLending.Models.Shared
         /// The bank or other financial institution providing the account.
         /// </summary>
         [JsonProperty("institution")]
-        public Dictionary<string, object> Institution { get; set; } = default!;
+        public AccountInstitution Institution { get; set; } = default!;
 
         [JsonProperty("modifiedDate")]
         public string? ModifiedDate { get; set; }