+++ lending/CodatLending/Models/Shared/BankingAccountBalance.cs
@@ -11,7 +11,6 @@
 namespace CodatLending.Models.Shared
 {
     using Newtonsoft.Json;
-    using System.Collections.Generic;
     
     /// <summary>
     /// The Banking Account Balances data type provides a list of balances for a bank account including end-of-day batch balance or running balances per transaction.<br/>
@@ -40,7 +39,7 @@ namespace CodatLending.Models.Shared
         /// Depending on the data provided by the underlying bank, not all balances are always available.
         /// </summary>
         [JsonProperty("balance")]
-        public Dictionary<string, object> Balance { get; set; } = default!;
+        public AccountBalanceAmounts Balance { get; set; } = default!;
 
         /// <summary>
         /// In Codat&apos;s data model, dates and times are represented using the &lt;a class=&quot;external&quot; href=&quot;https://en.wikipedia.org/wiki/ISO_8601&quot; target=&quot;_blank&quot;&gt;ISO 8601 standard&lt;/a&gt;. Date and time fields are formatted as strings; for example:<br/>